Scroll to the selected feed item in the list.
To be able to scroll the item in the list Angular provides $anchorScroll,
which watch $location.hash() to check for changes and scrolls to the item.
The problem is that Angular reloads the template when the hash changes, to
overcome this set "reloadOnSearch: false" on the route to work around it.